+++ Resources/SysPatch.py
@@ -214,7 +214,7 @@ class PatchSysVolume:
                 self.gpu_accel_patches_11()
                 rebuild_required = True
 
-        if self.model == "iMac7,1":
+        if self.model in ["iMac7,1", "iMac8,1"]:
             print("- Fixing Volume Control Support")
             self.add_new_binaries(ModelArray.AddVolumeControl, self.constants.audio_path)