+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 - Fix NightShift accidentally disabling on Minimal SMBIOS configs
 - Fix iMac9,1 audio support
 - Heavily expand Graphics ID list
-- Fix iMac7,1 audio support
+- Fix iMac7,1 and iMac8,1 audio support
 - Work-around Bluetooth Kernel Panic on Apple's Bluetooth 2.0 Controllers (USB 05AC:8206)
   - Affects iMac7,1 and MacPro3,1
 - Fix iMac external display support